<fix>[header]: fix useless shift
longs should not be shifted by zero or more than their number of bits-1

Related: ZSTAC-53781

+++ header/src/main/java/org/zstack/header/exception/SerialVersionUID.java
@@ -1,7 +1,7 @@
 package org.zstack.header.exception;
 
 public interface SerialVersionUID {
-    public static final long Base = 0x564D4F70 << 32;
+    public static final long Base = (long) 0x564D4F70 << 32;
 
     public static final long CloudRuntimeException = Base | 0x01;
     public static final long CloudConfigureFailException = Base | (0x01 << 1);
